improve static OpenGraph tags generation
