+++ packages/atlas-meta-server/package.json
@@ -1,5 +1,6 @@
 {
   "name": "@joystream/atlas-meta-server",
+  "version": "1.1.0",
   "description": "Node server for pre rendering html meta tags",
   "license": "GPL-3.0",
   "scripts": {
