+++ packages/atlas-meta-server/src/tags.ts
@@ -30,6 +30,34 @@ const sanitizeDescription = (appName: string, fullDescription?: string | null, t
   return needsTrimming ? oneLineDescription.slice(0, 157) + '...' : oneLineDescription
 }
 
+export const generateCommonMetaTags = (
+  appName: string,
+  appUrl: string,
+  title: string,
+  description: string,
+  image: string,
+  twitterId?: string | null
+): MetaTags => {
+  const sanitizedDescription = sanitizeDescription(appName, description, title)
+  return {
+    'og:title': title,
+    'twitter:title': title,
+    'description': sanitizedDescription,
+    'og:description': sanitizedDescription,
+    'twitter:description': sanitizedDescription,
+    'og:image': image,
+    'twitter:image': image,
+    'og:image:alt': title,
+    'twitter:image:alt': title,
+    'og:image:type': 'image/webp',
+    'og:site_name': appName,
+    'og:url': appUrl,
+    'og:type': 'website',
+    'twitter:card': 'summary_large_image',
+    ...(twitterId ? { 'twitter:site': twitterId } : {}),
+  }
+}
+
 export const generateVideoMetaTags = (
   video: BasicVideoFieldsFragment,
   thumbnailUrl: string,
@@ -39,20 +67,21 @@ export const generateVideoMetaTags = (
 ): MetaTags => {
   const videoUrl = joinUrlFragments(baseAppUrl, 'video', video.id)
   const videoEmbedUrl = joinUrlFragments(baseAppUrl, 'embedded', 'video', video.id)
-  const sanitizedDescription = sanitizeDescription(appName, video.description, video.title)
+  const commonTags = generateCommonMetaTags(
+    appName,
+    videoUrl,
+    video.title || 'Unnamed video',
+    video.description || '',
+    thumbnailUrl,
+    twitterId
+  )
 
   return {
-    ...(twitterId ? { 'twitter:site': twitterId } : {}),
-    'og:site_name': appName,
-    'og:title': video.title || '',
-    'og:description': sanitizedDescription,
+    ...commonTags,
     'og:type': 'video.other',
     'og:url': videoUrl,
-    'og:image': thumbnailUrl,
-    'og:image:alt': `Thumbnail for ${appName} video '${video.title}'`,
     'og:image:width': THUMBNAIL_WIDTH,
     'og:image:height': THUMBNAIL_HEIGHT,
-    'og:image:type': 'image/webp',
     'og:video': videoEmbedUrl,
     'og:video:secure_url': videoEmbedUrl,
     'og:video:width': video.mediaMetadata?.pixelWidth || VIDEO_WIDTH,
@@ -62,7 +91,7 @@ export const generateVideoMetaTags = (
     'twitter:player': videoEmbedUrl,
     'twitter:player:width': VIDEO_WIDTH,
     'twitter:player:height': VIDEO_HEIGHT,
-  }
+  } as const
 }
 
 export const generateChannelMetaTags = (
@@ -73,20 +102,20 @@ export const generateChannelMetaTags = (
   twitterId?: string
 ): MetaTags => {
   const channelUrl = joinUrlFragments(baseAppUrl, 'channel', channel.id)
-  const sanitizedDescription = sanitizeDescription(appName, channel.description, channel.title)
+  const commonTags = generateCommonMetaTags(
+    appName,
+    channelUrl,
+    channel.title || 'Unnamed channel',
+    channel.description || '',
+    avatarUrl,
+    twitterId
+  )
 
   return {
-    ...(twitterId ? { 'twitter:site': twitterId } : {}),
-    'og:site_name': appName,
-    'og:title': channel.title || '',
-    'og:description': sanitizedDescription,
+    ...commonTags,
     'og:type': 'profile',
-    'og:url': channelUrl,
-    'og:image': avatarUrl,
-    'og:image:alt': `Avatar photo for ${appName} channel '${channel.title}'`,
     'og:image:width': AVATAR_SIZE,
     'og:image:height': AVATAR_SIZE,
-    'og:image:type': 'image/webp',
     'twitter:card': 'summary',
   }
 }
@@ -248,8 +277,13 @@ export const generateChannelSchemaTagsHtml = (
   return htmlTags.join('\n')
 }
 
-export const generateMetaHtml = (tags: MetaTags) => {
+export const generateMetaHtml = (tags: MetaTags, addHelmetAttr = false) => {
   return Object.entries(tags)
-    .map(([name, content]) => `<meta name="${name}" content="${content}">`)
+    .map(
+      ([name, content]) =>
+        `<meta name="${name}" property="${name}" content="${content}" ${
+          addHelmetAttr ? 'data-react-helmet="true"' : ''
+        }>`
+    )
     .join('\n')
 }
