+++ packages/atlas/src/embedded/index.html
@@ -3,8 +3,9 @@
   <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <meta name="description" content="Watch, publish, sell and interact with videos on Joystream network." />
-    <title>%appName%</title>
+
+    <!--  The following node will be replaced by meta tags (including <title>) generated during Vite's build step (see ../plugins)  -->
+    <meta-tags />
 
     <!--  favicon and theme colors  -->
     <link rel="icon" href="/favicon.ico" sizes="any" />
