+++ packages/atlas/src/hooks/useHeadTags.tsx
@@ -1,4 +1,4 @@
-import { MetaTags } from '@joystream/atlas-meta-server/src/tags'
+import { MetaTags, generateCommonMetaTags } from '@joystream/atlas-meta-server/src/tags'
 import { useMemo } from 'react'
 import { Helmet } from 'react-helmet'
 
@@ -7,8 +7,17 @@ import { atlasConfig } from '@/config'
 export const useHeadTags = (title?: string | null, metaTagsMapping: MetaTags = {}) => {
   return useMemo(() => {
     const pageTitle = title ? `${title} - ${atlasConfig.general.appName}` : atlasConfig.general.appName
-    const metaTags = Object.entries(metaTagsMapping).map(([name, content]) => (
-      <meta name={name} content={content.toString()} key={name} />
+    const commonTags = generateCommonMetaTags(
+      atlasConfig.general.appName,
+      atlasConfig.general.appDescription,
+      pageTitle,
+      atlasConfig.general.appDescription,
+      atlasConfig.general.appOgImgPath,
+      atlasConfig.general.appTwitterId
+    )
+    const allMetaTagsMapping = { ...commonTags, ...metaTagsMapping }
+    const metaTags = Object.entries(allMetaTagsMapping).map(([name, content]) => (
+      <meta name={name} content={content.toString()} property={content.toString()} key={name} />
     ))
     return (
       <Helmet>
